Update Controllers require to correct path
The blog controller was looking in the wrong place for the SDK files.
Seems that the CS SDK was initially being placed inside AppBundle folder but was moved at some point - since the composer file wasn’t looking there. To fix that I use dirname to get the correct folder and updated the lib to src.